+++ packages/sitecore-jss-nextjs/src/middleware/personalize-middleware.ts
@@ -7,8 +7,9 @@ import {
   ExperienceParams,
   getPersonalizedRewrite,
 } from '@sitecore-jss/sitecore-jss/personalize';
-import { SiteResolver } from '@sitecore-jss/sitecore-jss/site';
+import { SiteInfo, SiteResolver } from '@sitecore-jss/sitecore-jss/site';
 import { debug, NativeDataFetcher } from '@sitecore-jss/sitecore-jss';
+import { PosResolver } from '@sitecore-jss/sitecore-jss/personalize';
 
 export type PersonalizeMiddlewareConfig = {
   /**
@@ -43,6 +44,13 @@ export type PersonalizeMiddlewareConfig = {
    * @default localhost
    */
   defaultHostname?: string;
+  /**
+   * function to resolve point of sale for a site
+   * @param {Siteinfo} site to get info from
+   * @param {string} language to get info for
+   * @returns point of sale value for site/language combination
+   */
+  getPointOfSale?: (site: SiteInfo, language: string) => string;
 };
 
 /**
@@ -235,9 +243,9 @@ export class PersonalizeMiddleware {
     // Execute targeted experience in CDP
     const { ua } = userAgent(req);
     const params = this.getExperienceParams(req);
-    const pointOfSale = site.pointOfSale
-      ? site.pointOfSale[language] || site.pointOfSale[site.language]
-      : '';
+    const pointOfSale = this.config.getPointOfSale
+      ? this.config.getPointOfSale(site, language)
+      : PosResolver.resolve(site, language);
     const variantId = await this.cdpService.executeExperience(
       personalizeInfo.contentId,
       browserId,