+++ packages/sitecore-jss/src/personalize/index.ts
@@ -1,4 +1,5 @@
 ﻿export { personalizeLayout } from './layout-personalizer';
+export { PosResolver } from './pos-resolver';
 export {
   GraphQLPersonalizeService,
   GraphQLPersonalizeServiceConfig,